# src/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Projeto Criado para o Supermarket",
    "version": "1.0.0",
    "title": "Supermarket",
    "termsOfService": "nullable",
    "contact": {
      "email": "matheuspinheiro@softeam.com.br"
    },
    "license": {
      "name": "2020 SofTeam",
      "url": "https://softeam.com.br/"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/forgotPassword": {
      "post": {
        "tags": [
          "forgotPassword"
        ],
        "summary": "Recuperação de senha",
        "description": "Rota na qual fará a recuperação de senha através do email",
        "operationId": "forgotPassword",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Dados utilizados para recuperação de senha",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sucesso na operação"
          },
          "400": {
            "description": "Não foi possível realizar a operação"
          },
          "500": {
            "description": "Error interno do servidor"
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
  },
...
}